# packages/contracts/openapi/agents.yaml
openapi: "3.1.0"
info:
  title: SIGFA Agents API
  version: "2.0.0"
  description: |
    Contrat agents & compétences SIGFA : profils agents, machine à états de disponibilité,
    statistiques et import CSV. Servi sous /api/v1.

    **Dépend de** : CONTRACT-001 (core.yaml) — tous les schémas transverses ($ref core.yaml).

    **Lien routage de file** : les compétences (services traitables) et langues de l'agent
    sont les données d'entrée du moteur de routage intelligent (API-004). Ce contrat expose
    les données ; la logique de routage est hors contrat.
servers:
  - url: /api/v1
    description: SIGFA v1
security:
  - BearerAuth: []
...
paths:
...
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Obtenir le profil d'un agent
      description: |
        Retourne le profil complet d'un agent : services traitables (multi-sélection,
        ids du catalogue), langues parlées (FR | EN), agences
        d'affectation et horaires de travail.

        Scope **agency** — AGENCY_DIRECTOR ou MANAGER minimum.
      x-tenant-scope: agency
      x-required-role: MANAGER
      parameters:
        - $ref: "./core.yaml#/components/parameters/IdParam"
      responses:
        "200":
          description: Profil de l'agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
              example:
                id: "55555555-5555-4555-a555-555555555555"
                email: "kofi@banque-ci.com"
                firstName: "Kofi"
                lastName: "Asante"
                role: AGENT
                bankId: "11111111-1111-4111-a111-111111111111"
                agencyId: "33333333-3333-4333-a333-333333333333"
                status: AVAILABLE
                languages: ["FR"]
                serviceIds: ["77777777-7777-4777-a777-777777777777", "88888888-8888-4888-a888-888888888888"]
                agencyIds: ["33333333-3333-4333-a333-333333333333"]
                workSchedule:
                  monday: { start: "08:00", end: "17:00" }
                  tuesday: { start: "08:00", end: "17:00" }
                createdAt: "2025-06-01T00:00:00Z"
        "400":
          $ref: "./core.yaml#/components/responses/BadRequest"
        "401":
          $ref: "./core.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./core.yaml#/components/responses/Forbidden"
        "404":
          $ref: "./core.yaml#/components/responses/NotFound"
        "409":
          $ref: "./core.yaml#/components/responses/Conflict"
        "422":
          $ref: "./core.yaml#/components/responses/UnprocessableEntity"
        "429":
          $ref: "./core.yaml#/components/responses/TooManyRequests"
        "500":
          $ref: "./core.yaml#/components/responses/InternalServerError"

    patch:
      tags: [Agents]
      summary: Mettre à jour le profil d'un agent
      description: |
        Mise à jour partielle du profil agent : services traitables (serviceIds),
        langues parlées (FR | EN), agences d'affectation (agencyIds)
        et horaires de travail (workSchedule).

        **Marquage conseiller (MODEL-CONTRACT-B, additif — D5)** : `isRelationshipManager`,
        `displayName` et `photoUrl` permettent de marquer/démarquer l'agent comme conseiller
        clientèle et de configurer son affichage public. Réservé à AGENCY_DIRECTOR+.

        Scope **agency** — AGENCY_DIRECTOR minimum pour modifier les compétences.
        L'agent lui-même (self) peut mettre à jour ses préférences de langue.

        **Lien API-004** : la mise à jour des serviceIds et languages est propagée au
        moteur de routage de file (API-004) pour le matching compétence + langue.
      x-tenant-scope: agency
      x-required-role: AGENCY_DIRECTOR
      parameters:
        - $ref: "./core.yaml#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentProfileRequest"
            example:
              languages: ["FR", "EN"]
              serviceIds: ["77777777-7777-4777-a777-777777777777", "99999999-9999-4999-a999-999999999999"]
              workSchedule:
                monday: { start: "09:00", end: "18:00" }
      responses:
        "200":
          description: Profil agent mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentProfile"
              example:
                id: "55555555-5555-4555-a555-555555555555"
                email: "kofi@banque-ci.com"
                firstName: "Kofi"
                lastName: "Asante"
                role: AGENT
                bankId: "11111111-1111-4111-a111-111111111111"
                agencyId: "33333333-3333-4333-a333-333333333333"
                status: AVAILABLE
                languages: ["FR", "EN"]
                serviceIds: ["77777777-7777-4777-a777-777777777777", "99999999-9999-4999-a999-999999999999"]
                agencyIds: ["33333333-3333-4333-a333-333333333333"]
                createdAt: "2025-06-01T00:00:00Z"
        "400":
          $ref: "./core.yaml#/components/responses/BadRequest"
        "401":
          $ref: "./core.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./core.yaml#/components/responses/Forbidden"
        "404":
          $ref: "./core.yaml#/components/responses/NotFound"
        "409":
          $ref: "./core.yaml#/components/responses/Conflict"
        "422":
          $ref: "./core.yaml#/components/responses/UnprocessableEntity"
        "429":
          $ref: "./core.yaml#/components/responses/TooManyRequests"
        "500":
          $ref: "./core.yaml#/components/responses/InternalServerError"

  # ─── STATUT TEMPS RÉEL ────────────────────────────────────────────────────────
  /agents/{id}/status:
    post:
      tags: [Agents]
      summary: Changer le statut temps réel d'un agent
      description: |
        Met à jour le statut de disponibilité de l'agent selon la machine à états :

        **Transitions légales** :
        - AVAILABLE → SERVING (prise en charge d'un ticket)
        - SERVING → AVAILABLE (ticket clôturé)
        - AVAILABLE → PAUSED (pause)
        - PAUSED → AVAILABLE (retour de pause)
        - AVAILABLE → ABSENT (absence planifiée)
        - ABSENT → AVAILABLE (retour d'absence)
        - AVAILABLE → OFFLINE (fin de session)
        - OFFLINE → AVAILABLE (début de session)

        **Transition illégale** : si la transition demandée n'est pas légale (ex.
        SERVING → ABSENT avec ticket ouvert sans transfert préalable), la réponse
        est **409 `ILLEGAL_AGENT_TRANSITION`** avec le statut actuel et la transition
        demandée dans `details`.

        Scope **agency** — l'agent lui-même (self) ou MANAGER+.
      x-tenant-scope: agency
      x-required-role: AGENT
      parameters:
        - $ref: "./core.yaml#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentStatusRequest"
            example:
              status: PAUSED
              reason: "Pause déjeuner"
      responses:
        "200":
          description: Statut agent mis à jour
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatusResponse"
              example:
                id: "55555555-5555-4555-a555-555555555555"
                status: PAUSED
                previousStatus: AVAILABLE
                updatedAt: "2026-07-11T12:00:00Z"
        "400":
          $ref: "./core.yaml#/components/responses/BadRequest"
        "401":
          $ref: "./core.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./core.yaml#/components/responses/Forbidden"
        "404":
          $ref: "./core.yaml#/components/responses/NotFound"
        "409":
          description: |
            Transition illégale — la transition demandée n'est pas autorisée depuis
            le statut actuel (ex. SERVING → ABSENT avec ticket ouvert).
          content:
            application/json:
              schema:
                $ref: "./core.yaml#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ILLEGAL_AGENT_TRANSITION
                  message: "Transition illégale : un ticket est en cours de traitement. Transférez ou clôturez le ticket avant de passer en ABSENT."
                  details:
                    currentStatus: SERVING
                    requestedStatus: ABSENT
                    activeTicketId: "ffffffff-ffff-4fff-afff-ffffffffffff"
        "422":
          $ref: "./core.yaml#/components/responses/UnprocessableEntity"
        "429":
          $ref: "./core.yaml#/components/responses/TooManyRequests"
        "500":
          $ref: "./core.yaml#/components/responses/InternalServerError"

  # ─── STATISTIQUES AGENT ───────────────────────────────────────────────────────
  /agents/{id}/stats:
    get:
      tags: [Agents]
      summary: Statistiques de performance d'un agent
      description: |
        Retourne les statistiques de l'agent pour la période demandée :
        - **ticketsHandled** : nombre de tickets traités sur la période
        - **avgHandlingTime** : temps moyen de traitement (TMT moyen) en secondes sur le jour
        - **currentTicket** : ticket en cours de traitement (numéro + durée en secondes)

        Ces schémas sont conçus pour être réutilisés par le reporting (CONTRACT-006).

        **Règle "self"** : l'agent peut consulter ses propres statistiques (self).
        Les MANAGER+ peuvent consulter les stats de n'importe quel agent de leur agence.
      x-tenant-scope: agency
      x-required-role: AGENT
      parameters:
        - $ref: "./core.yaml#/components/parameters/IdParam"
        - name: period
          in: query
          required: false
          description: "Période d'agrégation des statistiques (day | week | month). Défaut : day."
          schema:
            type: string
            enum:
              - day
              - week
              - month
            default: day
      responses:
        "200":
          description: Statistiques de l'agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStats"
              example:
                agentId: "55555555-5555-4555-a555-555555555555"
                period: day
                ticketsHandled: 24
                avgHandlingTime: 480
                currentTicket:
                  ticketId: "ffffffff-ffff-4fff-afff-ffffffffffff"
                  number: "A042"
                  durationSeconds: 120
        "400":
          $ref: "./core.yaml#/components/responses/BadRequest"
        "401":
          $ref: "./core.yaml#/components/responses/Unauthorized"
        "403":
          $ref: "./core.yaml#/components/responses/Forbidden"
        "404":
          $ref: "./core.yaml#/components/responses/NotFound"
        "409":
          $ref: "./core.yaml#/components/responses/Conflict"
        "422":
          $ref: "./core.yaml#/components/responses/UnprocessableEntity"
        "429":
          $ref: "./core.yaml#/components/responses/TooManyRequests"
        "500":
          $ref: "./core.yaml#/components/responses/InternalServerError"

  # ─── IMPORT CSV ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (durée de vie : 15 minutes).
        Obtenu via POST /auth/login ou POST /auth/refresh (core.yaml).
  schemas:
...
    AgentLanguage:
      type: string
      description: |
        Langues parlées par un agent de guichet SIGFA.
        Utilisées par le moteur de routage de file (API-004) pour le matching
        compétence + langue lors de l'attribution des tickets.
        - FR : Français
        - EN : Anglais

        Décision PO 2026-07 : DIOULA et BAOULE retirés du périmètre
        (breaking change assumé — voir CHANGELOG du package contracts).
      enum:
        - FR
        - EN

    # ── Machine à états de disponibilité de l'agent ────────────────────────
...
    AgentProfile:
      type: object
      additionalProperties: false
      description: |
        Profil complet d'un agent de guichet : identité, compétences (services
        traitables), langues parlées, agences d'affectation et horaires.
        Les serviceIds et languages sont transmis au moteur de routage (API-004).
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - bankId
        - agencyId
        - status
        - languages
        - serviceIds
        - agencyIds
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identifiant unique de l'agent
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          $ref: "./core.yaml#/components/schemas/Role"
        bankId:
          type: string
          format: uuid
        agencyId:
          type: string
          format: uuid
          description: Agence principale d'affectation
        status:
          $ref: "#/components/schemas/AgentStatus"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/AgentLanguage"
          minItems: 1
          description: Langues parlées par l'agent (au moins une)
          example: ["FR", "EN"]
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
          description: Services traitables (ids du catalogue, multi-sélection)
        agencyIds:
          type: array
          items:
            type: string
            format: uuid
          description: Agences d'affectation (principale + secondaires)
        workSchedule:
          $ref: "#/components/schemas/WorkSchedule"
        phoneMasked:
          type: string
          description: "Numéro de téléphone masqué (conformité UEMOA — ex: +225 07 *** ** 01)"
          example: "+225 07 *** ** 01"
        isRelationshipManager:
          type: boolean
          description: |
            Conseiller clientèle (MODEL-CONTRACT-B, additif — D5). Si `true`, l'agent apparaît
            dans la liste publique nominative `GET /public/agencies/{agencyId}/relationship-managers`
            et peut être ciblé par un ticket (`targetManagerId`). Défaut : false.
          default: false
        displayName:
          type: string
          description: |
            Nom d'affichage public du conseiller (MODEL-CONTRACT-B, additif — D5). Seul champ
            nominatif exposé publiquement. Aucun lien client↔conseiller attitré (hors-scope CRM).
          example: "Kofi A."
        photoUrl:
          type: ["string", "null"]
          format: uri
          description: URL de la photo du conseiller (optionnel, MODEL-CONTRACT-B).
        createdAt:
          type: string
          format: date-time

    # ── Requête de mise à jour du profil ──────────────────────────────────
    UpdateAgentProfileRequest:
      type: object
      description: Mise à jour partielle du profil agent
      additionalProperties: false
      properties:
        languages:
          type: array
          items:
            $ref: "#/components/schemas/AgentLanguage"
          minItems: 1
          description: Langues parlées (au moins une)
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
          description: Services traitables (ids du catalogue)
        agencyIds:
          type: array
          items:
            type: string
            format: uuid
          description: Agences d'affectation
        workSchedule:
          $ref: "#/components/schemas/WorkSchedule"
        phoneMasked:
          type: string
          description: Numéro de téléphone masqué (conformité UEMOA)
        isRelationshipManager:
          type: boolean
          description: |
            Marquer/démarquer l'agent comme conseiller clientèle (MODEL-CONTRACT-B, additif — D5).
            Réservé à AGENCY_DIRECTOR+ (RBAC de la route PATCH /agents/{id}).
        displayName:
          type: string
          description: Nom d'affichage public du conseiller (MODEL-CONTRACT-B, additif — D5).
          example: "Kofi A."
        photoUrl:
          type: ["string", "null"]
          format: uri
          description: URL de la photo du conseiller (optionnel, MODEL-CONTRACT-B).

    # ── Requête de changement de statut ───────────────────────────────────
    UpdateAgentStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/AgentStatus"
        reason:
          type: ["string", "null"]
          maxLength: 255
          description: Raison optionnelle du changement de statut (ex. "Pause déjeuner")
          example: "Pause déjeuner"

    # ── Réponse au changement de statut ───────────────────────────────────
    AgentStatusResponse:
      type: object
      required:
        - id
        - status
        - previousStatus
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/AgentStatus"
        previousStatus:
          $ref: "#/components/schemas/AgentStatus"
        updatedAt:
          type: string
          format: date-time

    # ── Ticket en cours (pour les stats) ──────────────────────────────────
...
    AgentStats:
      type: object
      description: |
        Statistiques de performance de l'agent pour la période demandée.
        Schémas réutilisables par le reporting (CONTRACT-006 les référencera).
      required:
        - agentId
        - period
        - ticketsHandled
        - avgHandlingTime
      properties:
        agentId:
          type: string
          format: uuid
        period:
          type: string
          enum:
            - day
            - week
            - month
          description: Période d'agrégation
        ticketsHandled:
          type: integer
          minimum: 0
          description: Nombre de tickets traités sur la période
        avgHandlingTime:
          type: integer
          minimum: 0
          description: Temps moyen de traitement (TMT moyen) en secondes sur le jour
        currentTicket:
          $ref: "#/components/schemas/CurrentTicketSummary"

    # ── Erreur d'import CSV (ligne) ────────────────────────────────────────
...